Upgrade Android Gradle Plugin and related dependencies
This commit updates the Android Gradle Plugin to version 8.11.1 and Kotlin to version 2.2.0.
It also updates the compile SDK version to 35 and Java compatibility to version 17.
Additionally, the AndroidManifest.xml has been updated to remove the package attribute as it's now inferred from the namespace in build.gradle.
Flutter and Dart SDK constraints in pubspec.yaml have also been updated.